+++ public/css/equipaments.css
@@ -1,0 +1,3 @@
+#body-panel {
+    padding: 2rem;
+}
