+++ src/controllers/clientsController.ts
@@ -1,0 +1,11 @@
+import { Request, Response } from 'express';
+import path from 'path';
+import { selectedMenu } from '../helpers/menuHelper';
+
+export const clients = (req: Request, res: Response) => {
+    res.render(path.join(__dirname, '../views/pages/clients.ejs'), {
+        pageName: 'Clientes',
+        menu: selectedMenu('clients')
+    });
+   
+}
