+++ src/controllers/equipamentsController.ts
@@ -1,0 +1,10 @@
+import { Request, Response } from 'express';
+import path from 'path';
+import { selectedMenu } from '../helpers/menuHelper';
+
+export const equipaments = (req: Request, res: Response) => {
+    res.render(path.join(__dirname, '../views/pages/equipaments.ejs'), {
+        pageName: 'Equipamentos',
+        menu: selectedMenu('equipaments')
+    });
+}
