+++ src/views/pages/clients.ejs
@@ -1,4 +1,6 @@
 <%- include('../partials/head.ejs') -%>
+<link rel="stylesheet" href="css/cliente.css">
+</head>
 <body>
     <%- include('../partials/aside-menu.ejs') -%>
