+++ src/views/pages/index.ejs
@@ -1,4 +1,6 @@
 <%- include('../partials/head.ejs') %>
+<link rel="stylesheet" href="css/home.css">
+</head>
 <body>
     <%- include('../partials/aside-menu.ejs') %>
     <div id="menu-container">
