+++ src/views/partials/aside-menu.ejs
@@ -1,9 +1,6 @@
 <aside class="aside-full">
     <div class="header">
         <h2>YEEZY</h2>
-        <div class="close-icon">
-            <img src="icons/close-menu.svg">
-        </div>
     </div>
     <div class="aside-buttons">
         <button class="btn-aside <% if(menu.home) {%> active-menu <% } %>">
@@ -35,6 +32,9 @@
         </button>
     </div>
     <div class="aside-github">
+        <div class="close-icon">
+            <img src="icons/close-menu.svg">
+        </div>
         <a href="https://github.com/AllysomFernando/YeezyManegement" target="_blank">
             <img src="icons/github.svg" alt="Github"/>
         </a>
