+++ src/views/partials/head.ejs
@@ -8,6 +8,3 @@
     <title><%- pageName -%> | Yeezy</title>
     <link rel="stylesheet" href="css/menu-lateral.css">
     <link rel="stylesheet" href="css/body.css">
-    <link rel="stylesheet" href="css/home.css">
-    <link rel="stylesheet" href="css/cliente.css">
-</head>
